refactor: make controller as interface
