+++ src/main/java/model/User.java
@@ -8,7 +8,7 @@ public class User {
     private final String name;
     private final String email;
 
-    public User(String userId, String password, String name, String email) {
+    public User(final String userId, final String password, final String name, final String email) {
         this.userId = userId;
         this.password = password;
         this.name = name;
